CD-27 Create an authorization middleware
